Improved APU ECAM visuals
Made colours of APU ECAM more lifelike. Colours of values change based on the value. If the APU voltage is above 110 and below 120 then the APU voltage value is green, otherwise it is amber. If the APU frequency is above 390 and below 410 then the APU frequency value is green, otherwise it is amber. If both of these are true, then the APUGenTitle is white, otherwise, it is amber.